+++ tests/test_utilities.py
@@ -34,8 +34,9 @@ def test_valid_filenames():
 def test_get_fully_qualified_name():
     console.banner("Fully Qualified Names")
     pairs = {
-        "py_trees.behaviour.Behaviour": py_trees.behaviour.Behaviour(),
-        "py_trees.decorators.Inverter": py_trees.decorators.Inverter(child=py_trees.behaviours.Success())
+        "py_trees.behaviours.Periodic": py_trees.behaviours.Periodic(name="foo", n=3),
+        "py_trees.decorators.Inverter": py_trees.decorators.Inverter(child=py_trees.behaviours.Success()),
+        "py_trees.behaviours.Success": py_trees.behaviours.Success(),
     }
     print(console.green + "------------------ Assertions ------------------\n" + console.reset)
     for expected_name, object_instance in pairs.items():